move to /pkg/utils
